Convert tests to pytest and tox
