+++ tox.ini
@@ -1,0 +1,20 @@
+[tox]
+envlist = py27,py36
+
+[testenv]
+deps=pytest
+commands=pytest tests
+
+[testenv:check]
+skip_install = true
+deps = flake8
+commands =
+    flake8
+
+[testenv:codecov]
+deps =
+    codecov
+skip_install = true
+commands =
+    coverage xml --ignore-errors
+    codecov []
